pyproject.toml
[tool.poetry]
name = "hackaton"
version = "0.1.0"
description = ""
authors = ["diegulio <diego.machado@usach.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
judini = "^0.0.20"
langchain = "^0.0.340"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
ipykernel = "^6.26.0"
streamlit = "^1.28.2"
black = "^23.11.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
